% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesANOVA.R
\name{bayesANOVA}
\alias{bayesANOVA}
\title{Bayesian GASTURBINE ANOVA}
\usage{
bayesANOVA()
}
\value{
plots and MCMC diagnostics
}
\description{
Bayesian GASTURBINE ANOVA
}
\examples{
\dontrun{bayesANOVA()}
}
