% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccts.R
\docType{data}
\name{ccts}
\alias{ccts}
\title{Circuits}
\format{
\subsection{}{

A data frame with 11 rows and 3 columns:
\describe{
\item{CCTS}{CCT number}
\item{STANDARD}{0 to 1}
\item{HUFFMAN}{0 to 1}
...
}
}
}
\source{
MS file
}
\usage{
ccts
}
\description{
Data on CCTS
}
\keyword{datasets}
