% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylogistic.R
\name{mylogistic}
\alias{mylogistic}
\title{Logistic Regression}
\usage{
mylogistic(
  fname = "HtWtData110.csv",
  yname = "male",
  xname = "weight",
  savedsteps = 15000
)
}
\arguments{
\item{fname}{file name}

\item{yname}{y name}

\item{xname}{x name}

\item{savedsteps}{default 15000}
}
\value{
Jags output
}
\description{
This activates a JAGS script to analyze logistic data
}
\examples{
\dontrun{mylogistic()}

}
\references{
JK Chapter 21
}
\concept{logistic}
