% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaired.R
\name{mypaired}
\alias{mypaired}
\title{My Paired 2 Sample Problem}
\usage{
mypaired(Data)
}
\arguments{
\item{Data}{Data frame}
}
\value{
Jags output
}
\description{
This is a sample package to help in the making of functions using JAGS scripts
}
\examples{
\dontrun{mypaired(Data = ccts)}

}
