% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypaired.R
\name{mypaired}
\alias{mypaired}
\title{mypaired}
\usage{
mypaired(Data)
}
\arguments{
\item{Data}{data source}
}
\value{
JAGS stuff
}
\description{
A JAGS script for Bayesian analysis for paired data
}
\examples{
\dontrun{mypaird(df)}
}
